connections: BLEStreamConnection for PUPConnection
This reverses the inheritance structure of the PUPConnection. Instead of
extending a data parser with tranmission functionality, this builds the
PUPConnection on top of a fairly generic BLE tx/rx stream.

This swaps in the minimal changes to get it working as in the previous
state, but it needs to be cleaned up to make logging/printing consistent
again.