main: hacks to make macOS work
Bleak is not nice and starts running stuff on the asyncio event loop
as soon as it is imported. Calling asyncio.run creates a new run loop,
which breaks bleak. We fix this by running using the already existing
run loop. Also the not nice stuff is never removed from the run loop
so we have to remove it manually.