main: new ipykernel command
This command simplifies creating an ipykernel using python runtime
that invokes the command. This means that if pybricksdev in installed
using pipx, then the ipykernel will use the same virtualenv where
pybricksdev is installed. This is convenient but fragile if additional
packages are needed in the ipykernel. We would also want to include
things like matplotlib, which would bloat the dependencies.