dfu: add fallback to use dfu-util command line tool
Installing libusb on Windows and configuring it is a pain. It is much
easier to get dfu-util. This adds a fallback to use dfu-util if pydfu
fails.

Recommended solution for now is to download http://dfu-util.sourceforge.net/releases/dfu-util-0.9-win64.zip
and copy `dfu-util-static.exe` to `.venv\Scripts\` in the project directory.

Also fixed failure due to missing build directory by using temporary directory.
This way we don't leave files behind on the users computer.