dfu: embed windows dfu-util
It is a bit of a chore to get dfu-util for windows and add it to your
PATH. It is not very big, so instead of asking all users to do this, we
just embed the file.

Other OSes have package managers that make getting this tool installed
much easier.