connections: fix checksum race condition
This fixes a race condition where the nus_handler() could be called
multiple times before returning to send_block() when it is waiting for
the checksum_ready event.

By setting self.expected_checksum = -1 in the nus_handler() callback
itself, we prevent the next call from assuming that the data is also
a checksum.

Issue: https://github.com/pybricks/pybricksdev/issues/16