repl_installer: Drop use of aioserial.
Previously, we used this library to mimic Pybricks BLE behavior on a Prime Hub when it still only supported USB. This allowed us to download and run programs in the same way on all hubs. By now, that hub has BLE support so this USB hack is no longer needed.

For the REPL installer, we don't need such complexity, so we can use a simple serial. This makes it simpler and more cross platform.