flash: don't rely on delay for BLE writes
This removes the delay option for BLE writes and updates the flash
algorithm to be the same as Pybricks code.

The required delay can vary by one order of magnitude. Case in point,
my Windows laptop requires a delay of 50ms to flash successfully while
on a good day, my Linux desktop can flash with a delay of 5ms.