connections: add hack for mpy-cross + ipython + windows
Due to a combination of technical reasons, we can run subprocesses
asynchronously in ipython kernels. This adds a fallback for that case
where mpy-cross is just called synchronously. It should be fine as long
as no one uses it with `asyncio.gather()` in their notebook.

Issue: https://github.com/pybricks/support/issues/177