ble.lwp3: add bytecodes and messages modules
This adds two new modules. The `bytecodes` module contains enums and
a few other special types that give meaning to binary values. The
`messages` module formats the bytecodes into messages that hubs will
understand.

So far, the hub bytecodes and messages are complete. The port bytecodes
and messages are TODO.