cli: only allow file path in run and compile commands
This changes the run and compile commands to only accept a file name
as an argument. Before this, if a file name was misspelled, it was
treated as the body of a MicroPython script and compiled instead of
raising a file not found error.

Since we are using argparse.FileType, `-` is accepted as an argument
meaning read from standard input. This will allow the previous behavior
of passing a script without actually creating a file, albeit with a
different syntax, i.e. piping via shell.

Fixes: https://github.com/pybricks/support/issues/411