dfu: fix dfu-util --device args
It seems that only the first arg after the first "," in the dfu-util
--device arg is used to match attached devices. Since SPIKE Prime and
MINDSTORMS RI have different product IDs, we need to get the right ID
to make it work for both hubs. We can do this by scraping the ID from
the dfu-util --list command.

Tested with dfu-util 0.9 on Ubuntu 20.04.

Issue: https://github.com/pybricks/support/issues/420