connections: add race_disconnect() method
This adds a new method for racing other async calls against a disconnect
event. This is especially important in the case where we are using
async queues. Without this, we would get a deadlock if the device
disconnected when we were waiting on a queue because nothing would
ever be pushed to the queue after the disconnect.