pybricksdev/flash: drop support for firmware.bin
We were allowing firmware.zip files with a "ready-to-go" firmware.bin
file, but this file was modified just as the firmware-base.bin file,
so there is really no difference between the two. To simplify things,
we only use one file name, firmware-base.bin.

This also makes having a main.py file optional so that we don't break
compatibility with the firmware.zip generated for SPIKE hubs.

Typings are added for the data contained in the firmware.metadata.json
file to clarify what ships in the firmware.zip file and what is added.