compile: Experimental support for multi-file projects.
This makes use of recent firmware changes to allow the main script to
import modules in the same folder. The modules can also import other
modules.

To ensure a smooth transition until a firmware with this feature is
released, scripts without local imports will be downloaded in the old
format (a single raw mpy data blob).