firmware: move create_firmware_blob function
This moves the create_firmware function from the flash module to the
firmware module since is strictly deals with the firmware.zip file.

Also add the license text to the return value while we are making
breaking changes. This way all info in the firmware.zip file is
included in the return value.

Also add the firmware module to the docs.