pybricksdev.connections.pybricks.PybricksHub: fix connection state management
PybricksHub.disconnect_observable was one-use only since it used an
AsyncSubject. This caused the PybricksHub.race_disconnect method to
think that the hub was still disconnected after a reconnection.

To fix this, we replace PybricksHub.disconnect_observable and
PybricksHub.connected with a new PybricksHub.connection_state_observable
that uses a BehaviorSubject so that it can be updated across multiple
connect/disconnect events.

This also introduces a new pybricksdev.connections.ConnectionState enum
that give a more nuanced state indication rather than using a bool.

Fixes: https://github.com/pybricks/support/issues/971