pybricksdev.ble: filter name is None in find_device()
A recent change in Bleak made this problem more apparent since it
happens nearly all of the time on Windows now. It is possible to get
an advertisement callback before the SCAN_RSP has been received in
which case there will be no local name yet. So we need to ignore those
responses and wait for one that does have the local name to ensure that
the found device has a useful .name attribute.

Fixes: https://github.com/orgs/pybricks/discussions/1010