Move PybricksHub client creation into constructor
Moves client creation into the PybricksHub constructor. A
future change will move this into a subclass specific for
BLE or USB, so configuring this is more appropriately done
in the constructor instead of the connect method.

The disconnect logic is moved into a class method so that
it can be shared by both BLE and USB implementations.

Temporarily comments out a message that prints the device
name as this is no longer available in the connect method.