pybricksdev: use absolute imports
Change imports from relative to absolute. Quick fix tools in the IDE
will automatically generate absolute imports so if we stick with that
we can use the tools without having to manually fix the imports. Also
absolute imports can be easier to understand if you don't know the
folder structure by heart.