compile: separate project directory and file path in compile_file()
mpy-cross saves the file path as part of the binary file. To keep things
small and not leak host computer details, we want to only include the
portion of the path that is relative to the project directory.

Previously, this could only be done by changing the current working
directory to the project directory before calling compile_file(). Now,
that is no longer necessary and the project directory is passed as a
separate argument.

This fixes not being able to compile files piped from stdin when using
the cli tool.