pybricksdev: use explicit text encoding for text files
Ensure that text files are opened with UTF-8 encoding by default to
avoid issues with different default encodings. This also requires
updating some dependencies that also had issues with text encoding.

PYTHONWARNDEFAULTENCODING is set for tests to try to catch any future
issues with text encoding. And PYTHONDEVMODE is set for good measure.