pybricksdev.dfu: Fix not rebooting hub on Windows.
Fix not rebooting hub on Windows after flashing firmware by using
the dfu-util.exe fallback. The --reset option doesn't work on Windows
because WinUSB doesn't support it. Instead, we can use the --dfuse-address
option with ":leave" to exit DFU mode and reboot the device. This is
a bit simpler than creating a .dfu file anyway.

Fixes: https://github.com/pybricks/pybricksdev/issues/112