compile: New implementation of multi-file compilation using mpy-tool.
Add a new implementation of `compile_multi_file()` that uses `mpy-tool`
to find imports. This will work better than the previous implementation
that used Python's `ModuleFinder`, which had several shortcomings.

This is only implemented for MPY ABI v6 for now. To make it work for
older ABIs, we would need to get an older copy of `mpy-tool`, and
backport the json output feature to it.